requirements.txt
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-documentai==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
ipykernel==5.3.4
ipython==7.18.0
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
proto-plus==1.8.1
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
rsa==4.6
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
